Refactor Code and write Unit test
